requirements.txt
opencv-python~=4.5.1.48
dlib~=19.21.1
pillow~=8.1.2
numpy~=1.20.1
pyserial>=3.5